Refactoricé logica del nivel
